pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hayfind"
version = "0.1.0"
description = "Local semantic search service and CLI for text repositories"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [
  { name = "Chris Fonte" }
]
dependencies = [
  "chromadb>=0.5.0",
  "fastapi>=0.111.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.1",
  "typer>=0.12.3",
  "uvicorn>=0.30.0",
  "google-generativeai>=0.8.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.0"
]

[project.scripts]
hayfind = "hayfind.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
